+++ packages/react-native/src/icons/Obsidian/style.ts
@@ -1,0 +1,9 @@
+export const TITLE = 'Obsidian';
+export const COMBINE_TEXT_MULTIPLE = 0.8;
+export const COMBINE_SPACE_MULTIPLE = 0.2;
+export const COLOR_PRIMARY = '#A88BFA';
+
+// Avatar constants
+export const AVATAR_BACKGROUND = '#000';
+export const AVATAR_COLOR = COLOR_PRIMARY;
+export const AVATAR_ICON_MULTIPLE = 0.7;