+++ src/Apertis/style.ts
@@ -1,0 +1,9 @@
+export const TITLE = 'Apertis';
+export const COMBINE_TEXT_MULTIPLE = 0.9;
+export const COMBINE_SPACE_MULTIPLE = 0.3;
+export const COLOR_PRIMARY = '#0d9488';
+
+// Avatar constants
+export const AVATAR_BACKGROUND = '#000';
+export const AVATAR_COLOR = COLOR_PRIMARY;
+export const AVATAR_ICON_MULTIPLE = 0.7;