+++ src/AskVerdict/components/Mono.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import { memo } from 'react';
+
+import type { IconType } from '@/types';
+
+import { TITLE } from '../style';
+
+const Icon: IconType = memo(({ size = '1em', style, ...rest }) => {
+  return (
+    <svg
+      fill="currentColor"
+      fillRule="evenodd"
+      height={size}
+      style={{ flex: 'none', lineHeight: 1, ...style }}
+      viewBox="0 0 24 24"
+      width={size}
+      xmlns="http://www.w3.org/2000/svg"
+      {...rest}
+    >
+      <title>{TITLE}</title>
+      <path d="M12 3.42L4.5 7.97 12 12.47Z" fillOpacity="0.7" />
+      <path d="M12 3.42L19.5 7.97 12 12.47Z" />
+      <path d="M12 6.56L14.48 8.95 13.08 8.95 12 7.73 10.83 8.95 9.42 8.95Z" fillOpacity="0.52" />
+      <path d="M4.5 13.88L12 16.83 19.5 13.88 19.5 15.61 12 18.56 4.5 15.61Z" fillOpacity="0.82" />
+      <path d="M5.3 16.83L12 19.55 18.7 16.83 18.7 18.38 12 20.77 5.3 18.38Z" fillOpacity="0.62" />
+    </svg>
+  );
+});
+
+export default Icon;