+++ src/LangChain/style.ts
@@ -1,9 +1,9 @@
 export const TITLE = 'LangChain';
-export const COMBINE_TEXT_MULTIPLE = 0.5;
+export const COMBINE_TEXT_MULTIPLE = 1;
 export const COMBINE_SPACE_MULTIPLE = 0.2;
-export const COLOR_PRIMARY = '#1C3C3C';
+export const COLOR_PRIMARY = '#7FC8FF';
 
 // Avatar constants
-export const AVATAR_BACKGROUND = COLOR_PRIMARY;
-export const AVATAR_COLOR = '#fff';
-export const AVATAR_ICON_MULTIPLE = 0.75;
+export const AVATAR_BACKGROUND = '#000';
+export const AVATAR_COLOR = COLOR_PRIMARY;
+export const AVATAR_ICON_MULTIPLE = 0.65;