+++ packages/react-native/CHANGELOG.md
@@ -2,6 +2,31 @@
 
 # Changelog
 
+## [Version 2.5.0](https://github.com/lobehub/lobe-icons/compare/@lobehub/icons-rn@2.4.0...@lobehub/icons-rn@2.5.0)
+
+<sup>Released on **2026-03-27**</sup>
+
+#### ✨ Features
+
+- **misc**: Add GeminiCLI, Reve, PrunaAI, BriaAI, Obsidian.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- **misc**: Add GeminiCLI, Reve, PrunaAI, BriaAI, Obsidian ([3fd24bd](https://github.com/lobehub/lobe-icons/commit/3fd24bd))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
 ## [Version 2.4.0](https://github.com/lobehub/lobe-icons/compare/@lobehub/icons-rn@2.3.0...@lobehub/icons-rn@2.4.0)
 
 <sup>Released on **2026-03-01**</sup>